backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyYAML==6.0.2
watchfiles==1.1.0
python-multipart==0.0.20
sqlalchemy==2.0.36
python-slugify==8.0.4
